vanir_config: workaround for asshats in modules assigning fstrict-aliasing with violations
Soo.. apparently some of the modules contain aliasing violations but have fstrict-
aliasing set in the AOSP code.  since we're asigning werror-strict-aliasing it breaks
the build.  Lets preserve the AOSP flags just in case.

To do: since fstrict-aliasing is assigned with -O2 level opts we will need to spend
a more time on this at some point

Change-Id: I0de5936341bdbc2ec88de5e5f76995213e2dfb63